Fixed access violation on FileExists function Added file log on PrintLog function Done adding file log function

+++ src/FileTable.cpp
@@ -347,10 +347,11 @@ void CFileTable::RenameFile(const char *pathFrom, const char* pathTo)
 
 bool FileExists(const char *path)
 {
-    int handle;
+    intptr_t handle;
     struct _finddata_t fileinfo;
 
     handle = _findfirst(path, &fileinfo);
+
     _findclose(handle);
 
     return handle == -1 ? false : strcmp(fileinfo.name, strrchr(path, '\\') + 1) == 0;  //filename must match case
